Add files via upload
Program based on the old fashioned tv show "You bet Your Life".
These programs focus on the usage of signals, which, in this program are redirected to be controlled by the user. Through a FIFO two terminals connect and talk to each other through a FIFO. Contestant writes guess and Groucho reads guess. If it's correct then game is won and FIFO is closed. 
The groucho program reads through the letter even if said on a sentence.