Add files via upload
Program like Twitter.
server.c sets up a connection with a port number and hostname through socket programming and client.c program uses that port number then both terminals communicate with each other. There is a max of 3 messages to be shared and program also asks for confirmation. Socket stays up and listening and displays so through a check. Also signals are implemented to kill the programs while closing the socket as well.